services_data-digestion-engine_1753589831.471000/requirements.txt
functions-framework==3.4.0

google-cloud-aiplatform>=1.42.0
vertexai>=1.4.0
google-cloud-discoveryengine>=0.10.6

google-cloud-storage
google-cloud-firestore
google-cloud-bigquery

opencv-python-headless

pydantic>=2.0
google-auth

google-adk